+++ src/app/static/src/app/store/adr/actions.ts
@@ -4,8 +4,8 @@ import {api} from "../../apiService";
 import qs from "qs";
 import {ADRState} from "./adr";
 import {ADRMutation} from "./mutations";
-import {constructUploadFile, datasetFromMetadata} from "../../utils";
-import {Organization, UploadFile, Dict} from "../../types";
+import {datasetFromMetadata} from "../../utils";
+import {Organization} from "../../types";
 import {BaselineMutation} from "../baseline/mutations";
 
 export interface ADRActions {
@@ -15,8 +15,7 @@ export interface ADRActions {
     getDatasets: (store: ActionContext<ADRState, RootState>) => void;
     getSchemas: (store: ActionContext<ADRState, RootState>) => void;
     getUserCanUpload: (store: ActionContext<ADRState, RootState>) => void;
-    getUploadFiles: (store: ActionContext<ADRState, RootState>) => void;
-    uploadFilestoADR: (store: ActionContext<ADRState, RootState>, uploadFilesPayload: UploadFile[]) => void;
+    getAndSetDatasets: (store: ActionContext<ADRState, RootState>, selectedDatasetId: string) => void;
 }
 
 export const actions: ActionTree<ADRState, RootState> & ADRActions = {
@@ -63,34 +62,14 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
     },
 
     async getUserCanUpload(context) {
-        const {state, rootState, commit} = context;
+        const {rootState, dispatch, commit} = context;
         const selectedDataset = rootState.baseline.selectedDataset;
 
         if (selectedDataset) {
-            //For backward compatibility, we may have to regenerate the dataset metadata to provide the
-            //organisation id for projects which are reloaded
-            let selectedDatasetOrgId: string;
             if (!selectedDataset.organization) {
-                //We may also have to fetch the selected dataset metadata too, if not loaded during this session
-                let datasets = state.datasets;
-                if (!datasets.length) {
-                    await api(context)
-                        .ignoreErrors()
-                        .ignoreSuccess()
-                        .get(`/adr/datasets/${selectedDataset.id}`)
-                        .then((response) => {
-                            if (response) {
-                                datasets = [response.data];
-                            }
-                        });
-                }
-
-                const regenDataset = datasetFromMetadata(selectedDataset.id, datasets, state.schemas!);
-                commit(`baseline/${BaselineMutation.SetDataset}`, regenDataset, {root: true});
-                selectedDatasetOrgId = regenDataset.organization.id;
-            } else {
-                selectedDatasetOrgId = selectedDataset.organization.id;
+                await dispatch("getAndSetDatasets", selectedDataset.id);
             }
+            const selectedDatasetOrgId = rootState.baseline.selectedDataset!.organization.id
 
             await api(context)
                 .withError(ADRMutation.SetADRError)
@@ -106,71 +85,21 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
         }
     },
 
-    async getUploadFiles(context) {
-        const {state, rootState, commit} = context;
-        const selectedDataset = rootState.baseline.selectedDataset;
-        const project = rootState.projects.currentProject;
-
-        if (selectedDataset && project) {
-            context.commit({type: ADRMutation.SetADRError, payload: null});
+    async getAndSetDatasets(context, selectedDatasetId) {
+        const {state, commit} = context;
+        let datasets = state.datasets;
+        if (!datasets.length) {
             await api(context)
-                .withError(ADRMutation.SetADRError)
+                .ignoreErrors()
                 .ignoreSuccess()
-                .get(`/adr/datasets/${selectedDataset.id}`)
+                .get(`/adr/datasets/${selectedDatasetId}`)
                 .then((response) => {
                     if (response) {
-                        const metadata = response.data;
-                        const schemas = state.schemas!;
-
-                        const uploadFiles = {
-                            outputZip: constructUploadFile(
-                                metadata,
-                                0,
-                                schemas.outputZip,
-                                `${project.name}_naomi_outputs.zip`,
-                                "uploadFileOutputZip"),
-                            outputSummary: constructUploadFile(
-                                metadata,
-                                1,
-                                schemas.outputSummary,
-                                `${project.name}_naomi_summary.html`,
-                                "uploadFileOutputSummary")
-                        };
-
-                        commit({type: ADRMutation.SetUploadFiles, payload: uploadFiles});
+                        datasets = [response.data];
                     }
                 });
         }
-    },
-
-    async uploadFilestoADR(context, uploadFilesPayload) {
-        const {state, rootState, commit} = context;
-        const selectedDatasetId = rootState.baseline.selectedDataset?.id;
-        const modelCalibrateId = rootState.modelCalibrate.calibrateId;
-        commit({type: ADRMutation.ADRUploadStarted, payload: uploadFilesPayload.length});
-
-        for (let i = 0; i < uploadFilesPayload.length; i++) {
-            commit({type: ADRMutation.ADRUploadProgress, payload: i + 1});
-            const { resourceType, resourceFilename, resourceId } = uploadFilesPayload[i]
-
-            const requestParams: Dict<string> = {resourceFileName: resourceFilename}
-            if (resourceId){
-                requestParams["resourceId"] = resourceId
-            }
-
-            let apiRequest = api<ADRMutation, ADRMutation>(context)
-                                        .withError(ADRMutation.SetADRUploadError);
-            if  (i === uploadFilesPayload.length - 1) {
-                apiRequest = apiRequest.withSuccess(ADRMutation.ADRUploadCompleted);
-            } else {
-                apiRequest = apiRequest.ignoreSuccess();
-            }
-            const response = await apiRequest.postAndReturn(`/adr/datasets/${selectedDatasetId}/resource/${resourceType}/${modelCalibrateId}`,
-                qs.stringify(requestParams))
-            if (!response) {
-                break
-            }
-        }
+        const regenDataset = datasetFromMetadata(selectedDatasetId, datasets, state.schemas!);
+        commit(`baseline/${BaselineMutation.SetDataset}`, regenDataset, {root: true});
     }
 };
-