+++ src/app/static/src/tests/integration/adr-dataset.itest.ts
@@ -3,11 +3,11 @@ import {actions as surveyAndProgramActions} from "../../app/store/surveyAndProgr
 import {login, rootState} from "./integrationTest";
 import {BaselineMutation} from "../../app/store/baseline/mutations";
 import {actions as adrActions} from "../../app/store/adr/actions";
+import {actions as adrUploadActions} from "../../app/store/adrUpload/actions";
 import {SurveyAndProgramMutation} from "../../app/store/surveyAndProgram/mutations";
 import {getFormData} from "./helpers";
 import {ADRMutation} from "../../app/store/adr/mutations";
 import {UploadFile} from "../../app/types";
-import {mockADRState} from "../mocks";
 
 // this suite tests all endpoints that talk to the ADR
 // we put them in a suite of their own so that we can run
@@ -104,7 +104,11 @@ describe("ADR dataset-related actions", () => {
     });
 
     it("can get dataset details on get userCanUpload when selected dataset organisation is not set", async () => {
-        const commit = jest.fn();
+        const commit = jest.fn().mockImplementation((mutation, payload) => {
+            if (mutation === "baseline/SetDataset") {
+                root.baseline.selectedDataset = payload
+            }
+        });
 
         // 1. get datasets
         await adrActions.getDatasets({commit, rootState} as any);
@@ -115,7 +119,7 @@ describe("ADR dataset-related actions", () => {
         const dataset = datasets.find((dataset: any) => dataset.organization.name === "naomi-development-team");
         expect(dataset).not.toBeNull()
 
-        // 3. check can upload
+        // 3. update dataset's organisation
         commit.mockClear();
         const root = {
             ...rootState,
@@ -125,11 +129,15 @@ describe("ADR dataset-related actions", () => {
                 }
             }
         };
-        const adr = { schemas, datasets: [] }; //do not include datasets in adr state, the action will fetch them
-        await adrActions.getUserCanUpload({commit, rootState: root, state: adr} as any);
-        expect(commit.mock.calls[0][0]).toBe(`baseline/${BaselineMutation.SetDataset}`);
-        expect(commit.mock.calls[1][0].type).toBe(ADRMutation.SetUserCanUpload);
-        expect(commit.mock.calls[1][0].payload).toBe(true);
+        const adr = {schemas, datasets: []}; //do not include datasets in adr state, the action will fetch them
+        await adrActions.getAndSetDatasets({commit, rootState, state: adr} as any, dataset.id)
+
+        // 4. check can upload
+        commit.mockClear();
+        const dispatch = jest.fn();
+        await adrActions.getUserCanUpload({commit, rootState: root, state: adr, dispatch} as any);
+        expect(commit.mock.calls[0][0].type).toBe(ADRMutation.SetUserCanUpload);
+        expect(commit.mock.calls[0][0].payload).toBe(true);
     });
 
     it("can import PJNZ file", async () => {
@@ -208,19 +216,27 @@ describe("ADR dataset-related actions", () => {
     it("hits upload files to adr endpoint and gets appropriate error", async () => {
 
         const commit = jest.fn();
+        const dispatch = jest.fn();
         const root = {
             ...rootState,
+            adr: { schemas, datasets: [] },
             modelCalibrate: {calibrateId: "calId"},
             baseline: {
                 selectedDataset: {
                     id: "datasetId",
                     organization: {id: "organisationId"}
                 }
+            },
+            modelRun: {
+                result: {
+                    uploadMetadata: {
+                        outputSummary: {description: "summary"},
+                        outputZip: {description: "zip"}
+                    }
+                }
             }
         };
         
-        const adr = { schemas, datasets: [] };
-
         const uploadFilesPayload = [
             {
                 resourceType: "type1",
@@ -229,8 +245,7 @@ describe("ADR dataset-related actions", () => {
             }
         ] as UploadFile[]
 
-        await adrActions.uploadFilestoADR({commit, state: adr, rootState: root} as any,
-            uploadFilesPayload);
+        await adrUploadActions.uploadFilesToADR({commit, dispatch, rootState: root} as any, uploadFilesPayload);
 
         expect(commit.mock.calls.length).toBe(3);
         expect(commit.mock.calls[0][0]["type"]).toBe("ADRUploadStarted");
@@ -239,6 +254,9 @@ describe("ADR dataset-related actions", () => {
         expect(commit.mock.calls[1][0]["payload"]).toBe(1);
         expect(commit.mock.calls[2][0]["type"]).toBe("SetADRUploadError");
         expect(commit.mock.calls[2][0]["payload"]["error"]).toBe("OTHER_ERROR");
+        expect(dispatch.mock.calls.length).toBe(2);
+        expect(dispatch.mock.calls[0][0]).toBe("adr/getAndSetDatasets");
+        expect(dispatch.mock.calls[1][0]).toBe("getUploadFiles");
     }, 7000);
 
 });